Fix for issue 11098. Fix includes code to change the shareable flag to update the status when the scope is Unshareable. Also the deleteResource() and semaphore.release() should happen only when the specific resource handle is removed from the resources list in the pool's datastructure. Reviewed by : Jagadish Ramu, Jerome. Tests passed : QL, jdbc/connector devtests.
svn path=/trunk/; revision=34564


Former-commit-id: 5c5d242a481daabb78078fdd34fd3d380f2def3b

+++ v3/connectors/connectors-runtime/src/main/java/com/sun/enterprise/resource/pool/datastructure/ListDataStructure.java
@@ -162,14 +162,17 @@ public class ListDataStructure implements DataStructure {
      * @param resource ResourceHandle
      */
     public void removeResource(ResourceHandle resource) {
+        boolean removed = false;
         synchronized (resources) {
             synchronized (free) {
                 free.remove(resource);
-                resources.remove(resource);
+                removed = resources.remove(resource);
             }
         }
-        dynSemaphore.release();
-        handler.deleteResource(resource);
+        if(removed) {
+            dynSemaphore.release();
+            handler.deleteResource(resource);
+        }
     }
 
     /**
